Implemented user login, register and authentication

+++ src/main/kotlin/com/example/notes/controllers/NoteController.kt
@@ -32,6 +32,7 @@ class NoteController(private val noteRepository: NoteRepository) {
     fun save(
         @RequestBody body: NoteRequest
     ): NoteResponse {
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
         val note = noteRepository.save(
             Note(
                 id = body.id?.let { ObjectId(it) } ?: ObjectId.get(),
@@ -39,7 +40,7 @@ class NoteController(private val noteRepository: NoteRepository) {
                 content = body.content,
                 color = body.color,
                 createdAt = Instant.now(),
-                ownerId = ObjectId()
+                ownerId = ObjectId(ownerId)
             )
         )
 
@@ -47,9 +48,8 @@ class NoteController(private val noteRepository: NoteRepository) {
     }
 
     @GetMapping
-    fun findByOwnerId(
-        @RequestParam(required = true) ownerId: String
-    ): List<NoteResponse> {
+    fun findByOwnerId(): List<NoteResponse> {
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
         return noteRepository.findByOwnerId(ObjectId(ownerId)).map {
             it.toResponse()
         }
@@ -57,7 +57,13 @@ class NoteController(private val noteRepository: NoteRepository) {
 
     @DeleteMapping(path = ["{id}"])
     fun deleteById(@PathVariable id: String) {
-        noteRepository.deleteById(ObjectId(id))
+        val note = noteRepository.findById(ObjectId(id)).orElseThrow {
+            IllegalArgumentException("Note not found")
+        }
+        val ownerId = SecurityContextHolder.getContext().authentication.principal as String
+        if (note.ownerId.toHexString() == ownerId) {
+            noteRepository.deleteById(ObjectId(id))
+        }
     }
 
 }
